공통 ì예외 처리

+++ src/main/java/com/example/concurrencycontrolproject/domain/seat/service/ScheduledSeatService.java
@@ -8,6 +8,8 @@ import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Service;
 
 import com.example.concurrencycontrolproject.domain.seat.entity.ScheduledSeat;
+import com.example.concurrencycontrolproject.domain.seat.exception.ScheduledSeatErrorCode;
+import com.example.concurrencycontrolproject.domain.seat.exception.ScheduledSeatException;
 import com.example.concurrencycontrolproject.domain.seat.repository.ScheduledSeatRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -26,19 +28,22 @@ public class ScheduledSeatService {
 		List<String> keys = Collections.singletonList(redisKey);
 		Long result = redisTemplate.execute(redisScript, keys, userId.toString());
 
-		if (result != null && result == 1) {
+		if (result == 0 || result == null) {
 			// Redis에 예약 정보 저장
-			ScheduledSeat scheduledSeat = new ScheduledSeat(redisKey, scheduleId, seatId, true, userId);
-			scheduledSeatRepository.save(scheduledSeat);
-			return true;
+			throw new ScheduledSeatException(ScheduledSeatErrorCode.SEAT_ALREADY_RESERVED);
 		}
-		return false;
+		ScheduledSeat scheduledSeat = new ScheduledSeat(redisKey, scheduleId, seatId, true, userId);
+		scheduledSeatRepository.save(scheduledSeat);
+		return true;
 	}
 
 	// 예약 취소
 	public void cancelReservation(Long scheduleId, Long seatId) {
 		String redisKey = "scheduled_seat:" + scheduleId + ":" + seatId;
 
+		if (!scheduledSeatRepository.existsById(redisKey)) {
+			throw new ScheduledSeatException(ScheduledSeatErrorCode.SEAT_NOT_FOUND);
+		}
 		// Redis에서 예약 데이터 삭제
 		scheduledSeatRepository.deleteById(redisKey);
 		redisTemplate.delete(redisKey);
@@ -47,7 +52,8 @@ public class ScheduledSeatService {
 	// 예약 상태 조회
 	public ScheduledSeat getReservation(Long scheduleId, Long seatId) {
 		String redisKey = "scheduled_seat:" + scheduleId + ":" + seatId;
-		return scheduledSeatRepository.findById(redisKey).orElse(null);
+		return scheduledSeatRepository.findById(redisKey)
+			.orElseThrow(() -> new ScheduledSeatException(ScheduledSeatErrorCode.SEAT_NOT_FOUND));
 	}
 }
